requirements.txt
aiogram~=3.4.1
PyYAML~=6.0.1